fixed modal again
